(MAINT) Fixups for latest style and dependency versions
Rubocop moved a bunch of cops around into new namespaces, and new style
suggestions exist for which the code has been updated, mainly
automatically.

This also updates the dependencies to use the new centralised gems which
removes (or moves) a bunch of cruft.

The travis configuration is also updated to use compatible Ruby versions
and to run the unit tests under moultiple Ruby versions as well.